# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chess Alive API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Chess Alive Team",
      "url": "https://github.com/yourusername/chess-alive"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/games": {
      "post": {
        "tags": ["games"],
        "summary": "Create Game",
        "description": "Create a new game. Returns a share code that other players can use to join.",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Player session ID (auto-generated if not provided)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}": {
      "get": {
        "tags": ["games"],
        "summary": "Get Game",
        "description": "Get the current state of a game including board position, pieces with morale, and player information.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Game UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Game state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        }
      }
    },
    "/api/v1/games/join-by-code": {
      "post": {
        "tags": ["games"],
        "summary": "Join Game by Code",
        "description": "Join an existing game using its share code.",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined game successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid share code"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/join": {
      "post": {
        "tags": ["games"],
        "summary": "Join Game (by Game ID)",
        "description": "Alternative join endpoint using game ID.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined game successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/command": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Issue Command",
        "description": "Issue a move command to a piece. The piece may refuse based on its morale and the riskiness of the move.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move or not your turn"
          },
          "404": {
            "description": "Game or piece not found"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/persuade": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Persuade Piece",
        "description": "Submit a persuasion attempt when a piece refuses to move. The AI evaluates your argument.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersuasionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persuasion processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersuasionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/resign": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Resign Game",
        "description": "Resign from the current game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resignation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "result": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/draw-offer": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Offer Draw",
        "description": "Offer a draw to your opponent.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draw offered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/draw-respond": {
      "post": {
        "tags": ["game-actions"],
        "summary": "Respond to Draw",
        "description": "Accept or decline a draw offer.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawRespondRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response recorded"
          }
        }
      }
    },
    "/api/v1/games/{game_id}/moves": {
      "get": {
        "tags": ["games"],
        "summary": "Get Move History",
        "description": "Get the complete move history for a game.",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Move history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "moves": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Move"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGameRequest": {
        "type": "object",
        "properties": {
          "game_mode": {"type": "string", "enum": ["pvp", "pvai"], "default": "pvp"},
          "template": {"type": "string", "enum": ["classic", "power_chess", "leaper_madness", "hopper_havoc", "pawn_revolution"], "default": "classic"},
          "settings": {"$ref": "#/components/schemas/GameSettings"}
        }
      },
...
      "JoinGameRequest": {
        "type": "object",
        "required": ["share_code"],
        "properties": {
          "share_code": {"type": "string", "pattern": "^[A-Z0-9]{6,12}$"}
        }
      },
      "GameResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "status": {"type": "string", "enum": ["waiting", "active", "completed"]},
          "game_mode": {"type": "string"},
          "template": {"type": "string"},
          "share_code": {"type": "string"},
          "fen": {"type": "string"},
          "turn": {"type": "string", "enum": ["white", "black"]},
          "white_player": {"$ref": "#/components/schemas/PlayerSummary"},
          "black_player": {"$ref": "#/components/schemas/PlayerSummary"},
          "pieces": {"type": "array", "items": {"$ref": "#/components/schemas/PieceState"}},
          "settings": {"$ref": "#/components/schemas/GameSettings"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "CommandRequest": {
        "type": "object",
        "required": ["piece_id", "target_square"],
        "properties": {
          "piece_id": {"type": "string"},
          "target_square": {"type": "string", "pattern": "^[a-h][1-8]$"},
          "message": {"type": "string", "maxLength": 500}
        }
      },
      "CommandResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "move_executed": {"type": "boolean"},
          "piece_response": {"$ref": "#/components/schemas/PieceResponseData"},
          "board_state": {"$ref": "#/components/schemas/BoardState"},
          "morale_changes": {"type": "array", "items": {"$ref": "#/components/schemas/MoraleChange"}},
          "analysis": {"$ref": "#/components/schemas/MoveAnalysis"},
          "king_taunt": {"type": "string", "nullable": true}
        }
      },
...
      "PersuasionRequest": {
        "type": "object",
        "required": ["piece_id", "target_square", "argument"],
        "properties": {
          "piece_id": {"type": "string"},
          "target_square": {"type": "string", "pattern": "^[a-h][1-8]$"},
          "argument": {"type": "string", "maxLength": 500},
          "is_voice": {"type": "boolean", "default": false}
        }
      },
      "PersuasionResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "probability": {"type": "number"},
          "piece_response": {"type": "string"},
          "move_executed": {"type": "boolean"},
          "board_state": {"$ref": "#/components/schemas/BoardState"},
          "evaluation": {"$ref": "#/components/schemas/PersuasionEvaluation"}
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-ID"
      }
    }
  }
}